Make anonymization more useful
Previously, for every string it generated anonymized version, it generated
new random string. This lead to cases like:

select *
    from tablea as a, tableb as b
    where a.whatever = 'xxx' and b.sth = 'xxx'

to have different values instead of the same 'xxx'. New version of code
keeps track of generated anonymized versions, and doesn't generate new one
when not necessary.

I also took the liberty of refactoring a bit marking things for skipping
anonymization - I think that the current way is faster (less
substitutions and relying on special strings), and it should also be
easier to extend to add new classes of things to avoid anonymization of.