emove html_highlight_code() method to apply HTML formatting style at token level instead of full SQL content like before. You just have to call the $object->beautify() method, the HTML formatting will be called internally if you have set the output format to HTML with the call to $object->format('html'). This call is now done internally by pg_format when it is called with the -F html option or in CGI mode. This compatibility break justify a change of the major version number.
Note that this breaks backward compatibility if you are using your
own custom scripts. With a complete upgrade and the use of the
pg_format script and libraries everything should work as before.

There is also some improvement in the CGI interface where the
footer bar was overlapping on object on the left side of the
screen.

The CGI library also support a new internal parameter/option
$self->{ 'colorize' } that is used to disable HTML output formatting
to avoid applying colorization to error messages returned by the
Perl package pgFormatter::CGI.pm

Improve documentation of the Perl Package.

Add detection of user defined function in the SQL code. Until now
only PostgreSQL known function was highlighted. Thanks to Gajus
Kuizinas for the feature request.