Correctly format casts to quoted types
When encountering a construct like [ col::"text" ], the input is
tokenized into [ col:: ] [ "text" ], and then when those tokens are
printed the final cast is omitted from the first token and it ends up
as [ col "text" ], which produces a syntax error.  Fix this by
outputting the trailing "::" from the first token and omitting the
space between tokens when the previous token ends with "::".

Theoretically this could be fixed by changing the tokenization so that
[ col::"text" ] is generated as a single token, but that seemed like
it was much more likely to introduce bugs elsewhere or require
significant changes to the tokenizing logic.

The hash_func.sql had to be updated because it uses this construct.
The formatted code from before this change produces a syntax error in
my local Postgres (v12.4), so this shouldn't affect working code.